+++ app/src/main/java/com/example/sport_planet/data/response/chatting/ChattingMessageListResponse.kt
@@ -1,4 +1,4 @@
-package com.example.sport_planet.data.response
+package com.example.sport_planet.data.response.chatting
 
 import com.google.gson.annotations.SerializedName
