+++ app/src/main/java/com/example/sport_planet/remote/RemoteDataSource.kt
@@ -6,8 +6,13 @@ import com.example.sport_planet.data.request.MyViewEditRequest
 import com.example.sport_planet.data.response.*
 
 import com.beust.klaxon.JsonObject
+import com.example.sport_planet.data.response.chatting.CommonServerResponse
 import com.example.sport_planet.data.response.basic.ExerciseResponse
 import com.example.sport_planet.data.response.basic.RegionResponse
+import com.example.sport_planet.data.response.chatting.ChattingMessageListResponse
+import com.example.sport_planet.data.response.chatting.ChattingRoomListResponse
+import com.example.sport_planet.data.response.chatting.MakeChattingMessageReadResponse
+import com.example.sport_planet.data.response.chatting.MakeChattingRoomResponse
 import com.example.sport_planet.data.response.login.LoginResponse
 import com.example.sport_planet.data.response.login.SignUpResponse
 import com.example.sport_planet.data.response.mypage.*
