+++ app/src/main/res/layout/item_chat_bot_notice_message.xml
@@ -6,7 +6,7 @@
         <import type="com.example.sport_planet.util.Util" />
         <variable
             name="ItemChatBotNoticeMessage"
-            type="com.example.sport_planet.data.model.ChattingMessageModel" />
+            type="com.example.sport_planet.data.model.chatting.ChattingMessageModel" />
     </data>
     <LinearLayout
         android:layout_width="match_parent"
