+++ app/src/main/res/layout/item_chatting_room.xml
@@ -8,7 +8,7 @@
         <import type="com.example.sport_planet.util.Util" />
         <variable
             name="ItemChattingRoom"
-            type="com.example.sport_planet.data.response.ChattingRoomListResponse.Data" />
+            type="com.example.sport_planet.data.response.chatting.ChattingRoomListResponse.Data" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
