+++ app/src/main/res/layout/item_notice_message.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="ItemNoticeMessage"
-            type="com.example.sport_planet.data.model.ChattingMessageModel" />
+            type="com.example.sport_planet.data.model.chatting.ChattingMessageModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
